api-key-manager/requirements.txt
# Core FastAPI Dependencies
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
python-dotenv==1.1.1

# Production Server (for deployment)
gunicorn==23.0.0

# Virtual Staging Dependencies
websocket-client==1.8.0
Pillow==11.3.0
